examples/quickstart_tensorflow/pyproject.toml
[build-system]
requires = [
    "poetry==1.1.13",
]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "quickstart_tensorflow"
version = "0.1.0"
description = "Keras Federated Learning Quickstart with Flower"
authors = ["The Flower Authors <enquiries@flower.dev>"]

[tool.poetry.dependencies]
python = ">=3.8.2,<3.10"
flwr = "^0.18.0"
numpy = "^1.22.3"
# flwr = { path = "../../", develop = true }  # Development
tensorflow-cpu = "^2.6.2"
